# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AdBoard API",
    "version": "v1"
  },
  "paths": {
    "/Advert/search": {
      "post": {
        "tags": [
          "Advert"
        ],
        "summary": "Получить список объявлений",
        "requestBody": {
          "description": "Детали запроса",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvertSearchRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvertSearchRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AdvertSearchRequestDto"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "Запрос выполнен успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShortAdvertDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Advert/{id}": {
      "get": {
        "tags": [
          "Advert"
        ],
        "summary": "Получить объявление по id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор объявления",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "Запрос выполнен успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvertDto"
                }
              }
            }
          },
          "404": {
            "description": "Объявление с таким id не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Advert"
        ],
        "summary": "Обновляет объявление по модели",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор обновляемого объявления",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "CategoryId",
                  "Cost",
                  "Location",
                  "Name",
                  "Phone"
                ],
                "type": "object",
                "properties": {
                  "Name": {
                    "maxLength": 256,
                    "minLength": 4,
                    "type": "string",
                    "description": "Название."
                  },
                  "Description": {
                    "maxLength": 256,
                    "minLength": 0,
                    "type": "string",
                    "description": "Описание."
                  },
                  "Images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Изображения объявлений"
                  },
                  "Cost": {
                    "minimum": 0,
                    "type": "number",
                    "description": "Цена.",
                    "format": "double"
                  },
                  "Email": {
                    "type": "string",
                    "description": "Email.",
                    "format": "email"
                  },
                  "Phone": {
                    "type": "string",
                    "description": "Телефон.",
                    "format": "tel"
                  },
                  "Location": {
                    "maxLength": 256,
                    "minLength": 1,
                    "type": "string",
                    "description": "Геолокация объявления"
                  },
                  "CategoryId": {
                    "type": "string",
                    "description": "Id категории.",
                    "format": "uuid"
                  }
                }
              },
              "encoding": {
                "Name": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "Images": {
                  "style": "form"
                },
                "Cost": {
                  "style": "form"
                },
                "Email": {
                  "style": "form"
                },
                "Phone": {
                  "style": "form"
                },
                "Location": {
                  "style": "form"
                },
                "CategoryId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "Запрос успешно выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortAdvertDto"
                }
              }
            }
          },
          "400": {
            "description": "Модель невалидна",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Объявление не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Конфликт бизнес логики",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Advert"
        ],
        "summary": "Удаляет объявление по ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор объявления",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "204": {
            "description": "Удаление произведено успешно"
          },
          "403": {
            "description": "Нет доступа",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Объявление с таким ID не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
    "/Advert/{id}/Comments": {
      "get": {
        "tags": [
          "Advert"
        ],
        "summary": "Получить список комментариев объявления",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор объявления",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "Запрос выполнен успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShortCommentDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Advert/{id}/comments": {
      "post": {
        "tags": [
          "Advert"
        ],
        "summary": "Создаёт комментарий по модели в объявлении",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор объявления",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "Text"
                ],
                "type": "object",
                "properties": {
                  "Text": {
                    "maxLength": 1024,
                    "minLength": 0,
                    "type": "string",
                    "description": "Содержимое комментария"
                  },
                  "ParentId": {
                    "type": "string",
                    "description": "Родитель (если комментарий является ответом на другой комментарий)",
                    "format": "uuid"
                  }
                }
              },
              "encoding": {
                "Text": {
                  "style": "form"
                },
                "ParentId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "201": {
            "description": "Успешно создано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortCommentDto"
                }
              }
            }
          },
          "400": {
            "description": "Модель данных запроса невалидна",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Произошёл конфликт бизнес логики",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
    "/Advert": {
      "post": {
        "tags": [
          "Advert"
        ],
        "summary": "Создаёт объявление по модели",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "CategoryId",
                  "Cost",
                  "Location",
                  "Name",
                  "Phone"
                ],
                "type": "object",
                "properties": {
                  "Name": {
                    "maxLength": 256,
                    "minLength": 4,
                    "type": "string",
                    "description": "Название."
                  },
                  "Description": {
                    "maxLength": 256,
                    "minLength": 0,
                    "type": "string",
                    "description": "Описание."
                  },
                  "Images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Изображения объявлений"
                  },
                  "Cost": {
                    "minimum": 0,
                    "type": "number",
                    "description": "Цена.",
                    "format": "double"
                  },
                  "Email": {
                    "type": "string",
                    "description": "Email.",
                    "format": "email"
                  },
                  "Phone": {
                    "type": "string",
                    "description": "Телефон.",
                    "format": "tel"
                  },
                  "Location": {
                    "maxLength": 256,
                    "minLength": 1,
                    "type": "string",
                    "description": "Геолокация объявления"
                  },
                  "CategoryId": {
                    "type": "string",
                    "description": "Id категории.",
                    "format": "uuid"
                  }
                }
              },
              "encoding": {
                "Name": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "Images": {
                  "style": "form"
                },
                "Cost": {
                  "style": "form"
                },
                "Email": {
                  "style": "form"
                },
                "Phone": {
                  "style": "form"
                },
                "Location": {
                  "style": "form"
                },
                "CategoryId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "201": {
            "description": "Объявление успешно создано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortAdvertDto"
                }
              }
            }
          },
          "400": {
            "description": "Модель данных запроса невалидна",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Произошёл конфликт бизнес логики",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdvertDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Id.",
            "format": "uuid"
          },
          "user": {
            "$ref": "#/components/schemas/ShortUserDto"
          },
          "name": {
            "type": "string",
            "description": "Название.",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Описание.",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "description": "Признак актуальности объявления."
          },
          "imagesIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID изображений привязанных к объявлению",
            "nullable": true
          },
          "cost": {
            "type": "number",
            "description": "Цена.",
            "format": "double"
          },
          "email": {
            "type": "string",
            "description": "Email.",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "description": "Телефон.",
            "nullable": true
          },
          "location": {
            "type": "string",
            "description": "Геолокация объявления",
            "nullable": true
          },
          "created": {
            "type": "string",
            "description": "Дата создания.",
            "format": "date-time"
          },
          "category": {
            "$ref": "#/components/schemas/ShortCategoryDto"
          }
        },
        "additionalProperties": false,
        "description": "Объявление (подробное описание)"
      },
      "AdvertSearchRequestDto": {
        "type": "object",
        "properties": {
          "search": {
            "type": "string",
            "description": "Строка поиска.",
            "nullable": true
          },
          "showNonActive": {
            "type": "boolean",
            "description": "Показывать неактивные объявления",
            "nullable": true
          },
          "category": {
            "type": "string",
            "description": "Категория",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Модель запроса на поиск объявлений."
      },
...
      "ErrorDto": {
        "type": "object",
        "properties": {
          "errorCode": {
            "type": "string",
            "description": "Код ошибки.",
            "nullable": true
          },
          "userMessage": {
            "type": "string",
            "description": "Сообщение для пользователя.",
            "nullable": true
          },
          "internalErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDto"
            },
            "description": "Вложенные ошибки.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Модель ошибки."
      },
...
      "ShortAdvertDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Название",
            "nullable": true
          },
          "location": {
            "type": "string",
            "description": "Геолокация объявления",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "Дата создания.",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean",
            "description": "Признак актуальности объявления."
          },
          "imagesIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Изображения",
            "nullable": true
          },
          "cost": {
            "type": "number",
            "description": "Цена",
            "format": "double"
          }
        },
        "additionalProperties": false,
        "description": "Краткая информация об объявлении"
      },
      "ShortCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "description": "Родительская категория",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Название",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Краткое описание категории"
      },
      "ShortCommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор",
            "format": "uuid",
            "nullable": true
          },
          "text": {
            "type": "string",
            "description": "Содержимое комментария",
            "nullable": true
          },
          "created": {
            "type": "string",
            "description": "Дата создания",
            "format": "date-time"
          },
          "parentId": {
            "type": "string",
            "description": "Идентификатор родительского комментария",
            "format": "uuid",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/ShortUserDto"
          }
        },
        "additionalProperties": false,
        "description": "Сокращенная информация о комментарии"
      },
...
      "ShortUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Имя пользователя",
            "nullable": true
          },
          "login": {
            "type": "string",
            "description": "Логин",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Краткая информация о пользователе"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme.  \n                        Enter 'Bearer' [space] and then your token in the text input below.\n                        Example: 'Bearer secretKey'",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}